+++ app.js
@@ -3,12 +3,12 @@ const express = require('express');
 const path = require('path');
 const mongoose = require('mongoose');
 const ejsMate = require('ejs-mate');
-const { campgroundSchema, reviewSchema } = require('./schemas.js');
-const catchAsync = require('./utils/catchAsync');
+// const { campgroundSchema, reviewSchema } = require('./schemas.js');
+
 const ExpressError = require('./utils/ExpressError');
 const methodOverride = require('method-override');
-const Campground = require('./models/campground');
-const Review = require('./models/review');
+const campgrounds = require('./routes/campgrounds');
+const reviews = require('./routes/reviews');
 
 //DATABASE CONNECTION************************************
 mongoose.connect('mongodb://localhost:27017/camprr', {
@@ -34,110 +34,13 @@ app.set('views', path.join(__dirname, 'views'));
 app.use(express.urlencoded({ extended: true }));
 app.use(methodOverride('_method'));
 
-const validateCampground = (req, res, next) => {
-	const { error } = campgroundSchema.validate(req.body);
-	if (error) {
-		const msg = error.details.map((el) => el.message).join(',');
-		throw new ExpressError(msg, 400);
-	} else {
-		next();
-	}
-};
-
-const validateReview = (req, res, next) => {
-	const { error } = reviewSchema.validate(req.body);
-	if (error) {
-		const msg = error.details.map((el) => el.message).join(',');
-		throw new ExpressError(msg, 400);
-	} else {
-		next();
-	}
-};
-
+app.use('/campgrounds', campgrounds);
+app.use('/campgrounds/:id/reviews', reviews);
 //ROUTES*********************************************************
 //HOME
 app.get('/', (req, res) => {
 	res.render('home');
 });
-//ALL
-app.get('/campgrounds', async (req, res) => {
-	const campgrounds = await Campground.find({});
-	res.render('campgrounds/index', { campgrounds });
-});
-//NEW CAMPGROUNDS FORM
-app.get('/campgrounds/new', (req, res) => {
-	res.render('campgrounds/new');
-});
-//POST NEW CAMPGROUND
-app.post(
-	'/campgrounds',
-	validateCampground,
-	catchAsync(async (req, res) => {
-		// if (!req.body.campground) throw new ExpressError('Invalid Campground Data', 400);
-		const campground = new Campground(req.body.campground);
-		await campground.save();
-		res.redirect(`/campgrounds/${campground._id}`);
-	})
-);
-//SHOW CAMPGROUND
-app.get(
-	'/campgrounds/:id',
-	catchAsync(async (req, res) => {
-		const campground = await Campground.findById(req.params.id).populate('reviews');
-		res.render('campgrounds/show', { campground });
-	})
-);
-//EDIT CAMPGROUND FORM
-app.get(
-	'/campgrounds/:id/edit',
-	catchAsync(async (req, res) => {
-		const campground = await Campground.findById(req.params.id);
-		res.render('campgrounds/edit', { campground });
-	})
-);
-//UPDATE CAMPGROUND
-app.put(
-	'/campgrounds/:id',
-	validateCampground,
-	catchAsync(async (req, res) => {
-		const { id } = req.params;
-		const campground = await Campground.findByIdAndUpdate(id, { ...req.body.campground });
-		res.redirect(`/campgrounds/${campground._id}`);
-	})
-);
-//DELETE
-app.delete(
-	'/campgrounds/:id',
-	catchAsync(async (req, res) => {
-		const { id } = req.params;
-		await Campground.findByIdAndDelete(id);
-		res.redirect('/campgrounds');
-	})
-);
-
-//REVIEWS
-app.post(
-	'/campgrounds/:id/reviews',
-	validateReview,
-	catchAsync(async (req, res) => {
-		const campground = await Campground.findById(req.params.id);
-		const review = new Review(req.body.review);
-		campground.reviews.push(review);
-		await review.save();
-		await campground.save();
-		res.redirect(`/campgrounds/${campground._id}`);
-	})
-);
-
-app.delete(
-	'/campgrounds/:id/reviews/:reviewId',
-	catchAsync(async (req, res) => {
-		const { id, reviewId } = req.params;
-		await Campground.findByIdAndUpdate(id, { $pull: { reviews: reviewId } });
-		await Review.findByIdAndDelete(req.params.reviewId);
-		res.redirect(`/campgrounds/${id}`);
-	})
-);
 
 // ERROR HANDLING********************************************
 app.all('*', (req, res, next) => {
