+++ routes/campgrounds.js
@@ -1,9 +1,21 @@
 const express = require('express');
 const router = express.Router();
-const catchAsync = require('./utils/catchAsync');
-const ExpressError = require('./utils/ExpressError');
-const Campground = require('./models/campground');
+const catchAsync = require('../utils/catchAsync');
+const { campgroundSchema } = require('../schemas.js');
+const ExpressError = require('../utils/ExpressError');
+const Campground = require('../models/campground');
 
+const validateCampground = (req, res, next) => {
+	const { error } = campgroundSchema.validate(req.body);
+	if (error) {
+		const msg = error.details.map((el) => el.message).join(',');
+		throw new ExpressError(msg, 400);
+	} else {
+		next();
+	}
+};
+
+//ALL
 router.get('/', async (req, res) => {
 	const campgrounds = await Campground.find({});
 	res.render('campgrounds/index', { campgrounds });
